chore: Move, split and rename TxLayout files
