+++ source/components/Layout/index.tsx
@@ -1,4 +1,2 @@
 export * from './Layout';
 export * from './OnboardingLayout';
-export * from './TxConfirmLayout';
-export * from './TxLayout';
