+++ source/pages/Transactions/Fee.tsx
@@ -1,5 +1,5 @@
 import { Form, Input } from 'antd';
-import React, { useState, FC, useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 
 import {
   Layout,
@@ -8,17 +8,15 @@ import {
   Tooltip,
   Icon,
 } from 'components/index';
-import { useStore, useQueryData } from 'hooks/index';
+import { useStore } from 'hooks/index';
 import { getController } from 'utils/browser';
-
-import { TxConfirm } from './TxConfirmLayout';
 
 interface IFee {
   onFinish: (fee: number) => any;
   title: string;
 }
 
-const _Fee: FC<IFee> = ({ title, onFinish }) => {
+const Fee: React.FC<IFee> = ({ title, onFinish }) => {
   const { getRecommendedFee } = getController().wallet.account.sys.tx;
 
   const { activeNetwork } = useStore();
@@ -133,93 +131,4 @@ const _Fee: FC<IFee> = ({ title, onFinish }) => {
   );
 };
 
-const titleResolver = (txType: string) => {
-  switch (txType) {
-    case 'CreateToken':
-      return 'Create Token';
-
-    case 'CreateNFT':
-      return 'Create NFT';
-
-    case 'MintToken':
-      return 'Mint Token';
-
-    case 'MintNFT':
-      return 'Mint NFT';
-
-    case 'TransferToken':
-      return 'Transfer Token';
-
-    case 'UpdateToken':
-      return 'Update Token';
-
-    default:
-      throw new Error('Unknown transaction type');
-  }
-};
-
-const callbackResolver = (txType: string) => {
-  let callbackName;
-
-  switch (txType) {
-    case 'CreateToken':
-      callbackName = 'confirmTokenCreation';
-      break;
-
-    case 'CreateNFT':
-      callbackName = 'confirmNftCreation';
-      break;
-
-    case 'MintToken':
-      callbackName = 'confirmTokenMint';
-      break;
-
-    case 'MintNFT':
-      callbackName = 'confirmMintNFT';
-      break;
-
-    // TODO TransferToken
-    // case 'TransferToken':
-    //   callbackName = 'confirmAssetTransfer';
-    //   break;
-
-    case 'UpdateToken':
-      callbackName = 'confirmUpdateToken';
-      break;
-
-    default:
-      throw new Error('Unknown transaction type');
-  }
-
-  return getController().wallet.account.sys.tx[callbackName];
-};
-
-interface ITxLayout {
-  txType: string;
-}
-
-export const TxLayout: FC<ITxLayout> = ({ txType }) => {
-  const { host, ...transaction } = useQueryData();
-  const [hasFee, setHasFee] = useState(false);
-
-  const setFee = (fee: number) => {
-    transaction.fee = fee;
-    setHasFee(true);
-  };
-
-  const title = titleResolver(txType);
-  const callback = callbackResolver(txType);
-
-  if (!hasFee) return <_Fee title={title} onFinish={setFee} />;
-
-  return (
-    <Layout canGoBack={false} title={title}>
-      <TxConfirm
-        title={title}
-        txType={txType}
-        transaction={transaction}
-        callback={callback}
-      />
-    </Layout>
-  );
-};
+export default Fee;
